ランキングの作成
ランキングの作成をした。
Clearのゲームオーバー条件を変更した。

+++ Chicken_Lace/Ranking.h
@@ -3,7 +3,7 @@
 struct RankingData {
 	int no;
 	char name[5];
-	long score;
+	double score;
 };
 
 #define RANKCOUNT 7
